Исправляет отсутствие цены поездки
