+++ src/components/route-info.js
@@ -1,28 +1,8 @@
 import AbstractComponent from './abstract-component';
 import {getMarkupDate} from '../utils/common';
 
-const calculateTotalPrice = (points) => {
-  let totalPrice = 0;
-
-  points.forEach((point) => {
-    totalPrice += point.price;
-
-    point.options.forEach((option) => {
-      totalPrice += option.price;
-    });
-  });
-
-  return totalPrice;
-};
-
 const createTemplate = (points) => {
-  if (points.length === 0) {
-    return (
-      `<p class="trip-info__cost">
-        Total: &euro;&nbsp;<span class="trip-info__cost-value">0</span>
-      </p>`
-    );
-  } else {
+  if (points.length !== 0) {
     const startPoint = points[0];
     const endPoint = points[points.length - 1];
 
@@ -34,20 +14,16 @@ const createTemplate = (points) => {
     const startCity = startPoint.city;
     const endCity = endPoint.city;
 
-    const totalPrice = calculateTotalPrice(points);
-
     return (
       `<div class="trip-info__main">
         <h1 class="trip-info__title">${startCity} &mdash; ... &mdash; ${endCity}</h1>
   
         <p class="trip-info__dates">${startDate}&nbsp;&mdash;&nbsp${endDate}</p>
-      </div>
-
-      <p class="trip-info__cost">
-        Total: &euro;&nbsp;<span class="trip-info__cost-value">${totalPrice}</span>
-      </p>`
+      </div>`
     );
   }
+
+  return ``;
 };
 
 export default class RouteInfoComponent extends AbstractComponent {
